requirements.txt
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.4.2
nltk==3.8.1
tensorflow==2.16.1
ttkbootstrap==1.10.1
pillow==10.1.0
matplotlib==3.8.2
ttkthemes==3.2.2
scikeras==0.13.0
customtkinter==5.2.0
schedule==1.1.0